requirements.txt
nextcord==3.1.1
chat-exporter==3.0.1
bs4==0.0.2
python-dotenv==1.2.2
